requirements.txt
fastapi==0.111.0
google-cloud-translate==3.15.3